OpenCities dataset + Segformer works on simple sampled data

+++ src/losses/dice.py
@@ -94,7 +94,7 @@ class DiceLoss(_Loss):
                 )  # N,H*W -> N,H*W, C
                 y_true = y_true.permute(0, 2, 1) * mask.unsqueeze(1)  # N, C, H*W
             else:
-                y_true = F.one_hot(y_true, num_classes)  # N,H*W -> N,H*W, C
+                y_true = F.one_hot(y_true.to(torch.long), num_classes)  # N,H*W -> N,H*W, C
                 y_true = y_true.permute(0, 2, 1)  # N, C, H*W
 
         if self.mode == MULTILABEL_MODE:
