+++ src/losses/jaccard.py
@@ -78,7 +78,7 @@ class JaccardLoss(_Loss):
             y_true = y_true.view(bs, -1)
             y_pred = y_pred.view(bs, num_classes, -1)
 
-            y_true = F.one_hot(y_true, num_classes)  # N,H*W -> N,H*W, C
+            y_true = F.one_hot(y_true.to(torch.long), num_classes)  # N,H*W -> N,H*W, C
             y_true = y_true.permute(0, 2, 1)  # H, C, H*W
 
         if self.mode == MULTILABEL_MODE:
